Various changes  - Removed join_conditional  - moved is_connected and get_symbol to utils and finance  - many code cleanups/corrections
